feat: Add policy for article

+++ app/Http/Controllers/ArticleController.php
@@ -6,15 +6,16 @@ use App\Http\Requests\StoreArticleRequest;
 use App\Http\Requests\UpdateArticleRequest;
 use App\Http\Resources\ArticleResource;
 use App\Models\Article;
+use Illuminate\Http\Request;
 
 class ArticleController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Request $request)
     {
-        $articles = Article::all();
+        $articles = $request->user()->articles()->latest()->get();
         return ArticleResource::collection($articles);
     }
 
@@ -32,6 +33,8 @@ class ArticleController extends Controller
      */
     public function show(Article $article)
     {
+        $this->authorize('view', $article);
+
         return new ArticleResource($article);
     }
 
@@ -40,6 +43,7 @@ class ArticleController extends Controller
      */
     public function update(UpdateArticleRequest $request, Article $article)
     {
+        $this->authorize('update', $article);
         $article->update($request->validated());
         return new ArticleResource($article);
     }
@@ -49,6 +53,7 @@ class ArticleController extends Controller
      */
     public function destroy(Article $article)
     {
+        $this->authorize('delete', $article);
         $article->delete();
         return response()->noContent();
     }
